fechamento_nota para base de consulta em outros repositórios
