+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioFechamentoTurma.cs
@@ -1,0 +1,7 @@
+﻿namespace SME.SGP.Dominio.Interfaces
+{
+    public interface IRepositorioFechamentoTurma : IRepositorioBase<FechamentoTurma>
+    {
+        
+    }
+}
